requirements.txt
python-telegram-bot==20.6
openai==1.30.1
python-dotenv==1.0.1
PyMuPDF==1.24.4
gTTS==2.5.1
